Move tests of walking functions out into their own file.
